backend/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.31.0",
    "cryptography>=46.0.4",
    "fastapi>=0.128.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
